feat: VObject message on norme update

+++ arc-web/src/main/java/fr/insee/arc/web/gui/norme/service/ServiceViewJeuxDeReglesCopie.java
@@ -1,7 +1,6 @@
 package fr.insee.arc.web.gui.norme.service;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.Map;
 
 import org.apache.logging.log4j.LogManager;
@@ -9,11 +8,7 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.stereotype.Service;
 import org.springframework.ui.Model;
 
-import fr.insee.arc.core.dataobjects.ArcPreparedStatementBuilder;
-import fr.insee.arc.core.dataobjects.ViewEnum;
-import fr.insee.arc.utils.dao.UtilitaireDao;
 import fr.insee.arc.utils.exception.ArcException;
-import fr.insee.arc.web.gui.all.util.ConstanteBD;
 
 @Service
 public class ServiceViewJeuxDeReglesCopie extends InteractorNorme {
@@ -84,96 +79,21 @@ public class ServiceViewJeuxDeReglesCopie extends InteractorNorme {
 
 	public String copieJeuxDeRegles(Model model) {
 		loggerDispatcher.info("Mon action pour copier un jeu de règles", LOGGER);
+		
 		// le jeu de regle à copier
 		Map<String, ArrayList<String>> selectionOut = views.getViewJeuxDeRegles().mapContentSelected();
 		// le nouveau jeu de regle
 		Map<String, ArrayList<String>> selectionIn = views.getViewJeuxDeReglesCopie().mapContentSelected();
-		HashMap<String, String> type = views.getViewJeuxDeReglesCopie().mapHeadersType();
-		if (!selectionIn.isEmpty()) {
+		
+		if (!selectionOut.isEmpty() && !selectionIn.isEmpty()) {
 
-			// excute the copy
 			try {
-			// columns found in all rules tables
-			String inCommonColumns = new StringBuilder().append(ConstanteBD.ID_NORME.getValue())
-					.append("," + ConstanteBD.PERIODICITE.getValue()).append("," + ConstanteBD.VALIDITE_INF.getValue())
-					.append("," + ConstanteBD.VALIDITE_SUP.getValue()).append("," + ConstanteBD.VERSION.getValue())
-					.toString();
-
-			// specific columns = column of the table minus common tables minus id_regle
-			// (rules generated id)
-			ArcPreparedStatementBuilder getTableSpecificColumns = new ArcPreparedStatementBuilder();
-			getTableSpecificColumns.append("\n SELECT string_agg(column_name,',') ");
-			getTableSpecificColumns.append("\n FROM information_schema.columns c ");
-			getTableSpecificColumns.append("\n WHERE table_schema||'.'||table_name ="
-					+ getTableSpecificColumns.quoteText(this.getSelectedJeuDeRegle()));
-			getTableSpecificColumns.append("\n AND column_name NOT IN ");
-			getTableSpecificColumns.append(
-					"\n ('" + inCommonColumns.replace(",", "','") + "','" + ConstanteBD.ID_REGLE.getValue() + "') ");
-
-			String specificColumns = UtilitaireDao.get(0).getString(null, getTableSpecificColumns);
-
-			// Build the copy query
-			ArcPreparedStatementBuilder requete = new ArcPreparedStatementBuilder();
-
-			requete.append("INSERT INTO " + this.getSelectedJeuDeRegle() + " ");
-			requete.append("(");
-			requete.append(inCommonColumns + "," + specificColumns);
-			requete.append(")");
-
-			requete.append("\n SELECT ");
-			requete.append(String.join(",", requete.quoteText(selectionOut.get(ConstanteBD.ID_NORME.getValue()).get(0)),
-					requete.quoteText(selectionOut.get(ConstanteBD.PERIODICITE.getValue()).get(0)),
-					requete.quoteText(selectionOut.get(ConstanteBD.VALIDITE_INF.getValue()).get(0)) + "::date ",
-					requete.quoteText(selectionOut.get(ConstanteBD.VALIDITE_SUP.getValue()).get(0)) + "::date ",
-					requete.quoteText(selectionOut.get(ConstanteBD.VERSION.getValue()).get(0))));
-			requete.append("," + specificColumns);
-
-			requete.append(" FROM " + this.getSelectedJeuDeRegle() + "  ");
-
-			requete.append(" WHERE ");
-
-			requete.append(String.join(" AND ", //
-					// condition about id_norm
-					ConstanteBD.ID_NORME.getValue() + requete.sqlEqual(
-							selectionIn.get(ConstanteBD.ID_NORME.getValue()).get(0),
-							type.get(ConstanteBD.ID_NORME.getValue())),
-					ConstanteBD.PERIODICITE.getValue()
-							// condition about PERIODICITE
-							+ requete.sqlEqual(selectionIn.get(ConstanteBD.PERIODICITE.getValue()).get(0),
-									type.get(ConstanteBD.PERIODICITE.getValue())),
-					ConstanteBD.VALIDITE_INF.getValue()
-							// condition about VALIDITE_INF
-							+ requete.sqlEqual(selectionIn.get(ConstanteBD.VALIDITE_INF.getValue()).get(0),
-									type.get(ConstanteBD.VALIDITE_INF.getValue())),
-					ConstanteBD.VALIDITE_SUP.getValue()
-							// condition about VALIDITE_SUP
-							+ requete.sqlEqual(selectionIn.get(ConstanteBD.VALIDITE_SUP.getValue()).get(0),
-									type.get(ConstanteBD.VALIDITE_SUP.getValue())),
-					ConstanteBD.VERSION.getValue()
-							// condition about VERSION
-							+ requete.sqlEqual(selectionIn.get(ConstanteBD.VERSION.getValue()).get(0),
-									type.get(ConstanteBD.VERSION.getValue()))
-
-			));
-			requete.append(" order by " + ConstanteBD.ID_REGLE.getValue() + " ;");
-
-			// delete the current rules before the copy
-			if (this.getSelectedJeuDeRegle().equals("arc.ihm_chargement_regle")) {
-				emptyRuleTable(this.views.getViewJeuxDeRegles(), dataObjectService.getView(ViewEnum.IHM_CHARGEMENT_REGLE));
-			} else if (this.getSelectedJeuDeRegle().equals("arc.ihm_normage_regle")) {
-				emptyRuleTable(this.views.getViewJeuxDeRegles(), dataObjectService.getView(ViewEnum.IHM_NORMAGE_REGLE));
-			} else if (this.getSelectedJeuDeRegle().equals("arc.ihm_controle_regle")) {
-				emptyRuleTable(this.views.getViewJeuxDeRegles(), dataObjectService.getView(ViewEnum.IHM_CONTROLE_REGLE));
-			} else if (this.getSelectedJeuDeRegle().equals("arc.ihm_mapping_regle")) {
-				emptyRuleTable(this.views.getViewJeuxDeRegles(), dataObjectService.getView(ViewEnum.IHM_MAPPING_REGLE));
-			}
-
-				UtilitaireDao.get(0).executeRequest(null, requete);
-			} catch (ArcException ex) {
-				loggerDispatcher.error("Error in copieJeuxDeRegles", ex, LOGGER);
-			}
+			dao.execQueryCopieJeuxDeRegles(views.getViewJeuxDeRegles(), views.getViewJeuxDeReglesCopie(), this.getSelectedJeuDeRegle());
 			
 			this.vObjectService.destroy(views.getViewJeuxDeReglesCopie());
+		} catch (ArcException ex) {
+			loggerDispatcher.error("Error in copieJeuxDeRegles", ex, LOGGER);
+		}
 		} else {
 			loggerDispatcher.info("No rule set choosed", LOGGER);
 			this.views.getViewJeuxDeRegles().setMessage("normManagement.copyRuleset.noSelection");
